rename files to camel case
