+++ src/ddbToEs/ddbToEsSync.ts
